pyproject.toml
[tool.poetry]
name = "motionpatches"
version = "0.1.0"
description = "motionpatches"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
torch = {version = "2.0.1+cu118", source = "torch_cu118"}
torchvision = {version = "0.15.2+cu118", source = "torch_cu118"}
torchaudio = {version = "2.0.2+cu118", source = "torch_cu118"}
hydra-core = "1.3.2"
tqdm = "4.66.2"
transformers = "4.31.0"
opencv-python = "4.9.0.80"
einops = "0.7.0"
timm = "0.9.5"


[[tool.poetry.source]]
name = "torch_cu118"
url = "https://download.pytorch.org/whl/cu118"
priority = "explicit"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pysen]
version = "0.11"

[tool.pysen.lint]
enable_black = true
enable_flake8 = true
enable_isort = true
enable_mypy = true
mypy_preset = "strict"
line_length = 88
py_version = "py311"

[[tool.pysen.lint.mypy_targets]]
  paths = ["."]